@@ -1,24 +1,20 @@
-'use strict';
+import test from 'node:test';
+import assert from 'node:assert';
 
-/* eslint-env jest */
-
-const dual = require('../../lib/utils/dual.js');
+import * as dual from '../../src/utils/dual.js';
 
 const license = 'MIT';
 const dualLicense = '(MIT OR Apache-2.0)';
 
 test('Checks if is dual license or not.', () => {
-  expect.assertions(2);
-  expect(dual.isDual(dualLicense)).toBeTruthy();
-  expect(dual.isDual(license)).toBeFalsy();
+  assert.strictEqual(dual.isDual(dualLicense), true);
+  assert.strictEqual(dual.isDual(license), false);
 });
 
 test('Extracts the first part of the dual license string.', () => {
-  expect.assertions(1);
-  expect(dual.first(dualLicense)).toBe('MIT');
+  assert.strictEqual(dual.first(dualLicense), 'MIT');
 });
 
 test('Extracts the second part of the dual license string.', () => {
-  expect.assertions(1);
-  expect(dual.last(dualLicense)).toBe('Apache-2.0');
+  assert.strictEqual(dual.last(dualLicense), 'Apache-2.0');
 });
