@@ -3,7 +3,7 @@
 const js2xmlparser = require('js2xmlparser');
 
 const unifiedList = require('../utils/unified-list');
-const dual = require('../utils/dual');
+const dual = require('../../src/utils/dual');
 const project = require('../utils/project');
 const versionHandler = require('../utils/version-handler');
 const messages = require('../utils/messages');
