@@ -7,7 +7,7 @@ const consoleModule = require('./console.mdl');
 const project = require('../utils/project');
 const html = require('../utils/html');
 const messages = require('../utils/messages');
-const dual = require('../utils/dual');
+const dual = require('../../src/utils/dual');
 
 let ldir = '';
 
