@@ -1,6 +1,6 @@
 'use strict';
 
-const dual = require('./dual');
+const dual = require('../../src/utils/dual');
 const urlFetcher = require('./url-fetcher');
 const fileReader = require('./file-reader');
 const namesMap = new Map();
